Environments work, but only with manual specifying.
In work for issue #6 (add --confdir and remove the other parameters),
I've discovered what seems to be a bug with Puppet Master and how it
does things.

With a Puppet Agent with --environment production, I _have to_ specify
--environmentpath on the Puppet Master in order for it to pick up the
changes and attempt to run them. Even though I'm specifying --confdir
and it _should_ pick up $confdir/environments for --environmentpath, it
doesn't and requires manually specifying it on the command line.

I'm about to test and verify that this is actually what is going on.