Environments/manifest mode now possible.
Defined a new environment variable `PUPPETMASTER_ENVIRONMENTS_ENABLED`
which will switch the initial configuration of Puppet to environments
mode. By default, we'll default over to the deprecated manifests mode,
but when this variable is defined, we'll use Puppet directory
environments for nodes. This has been tested locally and it works for
both modes.

Removed the old Puppet Master startup script, as we're on NGINX now.