Better fix for #14 and 1.0.3, 1.0.4.
The previous fixes tried to address the problem with Puppet
configuration file changes. The underlying problem is that the `PATH`
environment variable is _absent_ when starting the Passenger process as
user `puppet` for the Puppet Master. By simply appending the `PATH`
variable to the environment variables in `config.ru`, everything just
works. No need for `path` or `certname` in Puppet configuration, just
workingness.